use hooks instead of hoc in ManagePricing
