+++ src/pages/ManagePricing/index.js
@@ -1,4 +1,3 @@
 import Component from './ManagePricing'
-import connector from './ManagePricing.connector'
 
-export default connector(Component)
+export default Component
